Create Ruslan
This project provides users with a simple and convenient task management system. The main features of the project include:

Adding, editing, and deleting tasks.

Tracking the status of tasks.

An easy and intuitive user interface.


The project is built using Python and the Flask framework. It is designed for small teams or individual users.

In the future, a mobile app and multi-user mode are planned to be added to the project.

For additional support or to contribute, please use GitHub Issues and Pull Requests